refactor: rename similar-year folder
