+++ src/contexts/movies/movies.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { MoviesRepository } from './repository/movie.repository';
-import { SimilarYearService } from '../../services/similarYear/similar-year.service';
+import { SimilarYearService } from '../../services/similar-year/similar-year.service';
 
 @Injectable()
 export class MoviesService {
